requirements.txt
flask==3.0.3
python-dotenv==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==3.1.1
sqlalchemy-orm==1.2.10
Flask-WTF==1.2.2
WTForms==3.2.1
Flask-CKEditor==1.0.0
Bootstrap-Flask==2.4.1